make mig 196 compatible with older mysql versions
With utf8mb4 a constraint of 255 results in a maximum possible primary key of 1020 bytes which is larger then the maximum 767 bytes in older MySQL versions. Changing the constraint to 191 results in a max. primary key of 764 bytes which will work with those ancient versions and 191 is still enough for this column. The id's in the cron table are text, why we can't use numerical ID's here