Implement contest srx/stx and comments in SimpleFLE
DF3CB has specified ',***' and '.***' for contest exchange-received
and sent, as well as '<...>' for comments.

Comments may contain spaces and are handled ahead of splitting rows by
spaces.

Additionally, you may enter an arbitrary ADIF-field(*) in the form of
'<tx_pwr: 50>' or '<notes: Uses a straight key>', as long as the key
matches '[a-z_]*:' and the is implemented in wavelog's 'import'
in the Logbook-Model.