fix QSO panel when using a date format with commas
Commas are not an allowed URI character, and so they need to be
replaced with a URI-safe stand in before being used. Of the default
URI-safe characters, only '~%:\' are not used yet. I chose to use '%'
as it seems the least likely to be used in any future date formats.

Closes #2343.